requirements.txt
Django>=4.3
djangorestframework
psycopg2-binary
djangorestframework-simplejwt
gunicorn>=20.1.0
python-decouple>=3.8
dj-database-url
drf-spectacular